organize reducers source code
