+++ src/inspect_ai/scorer/_reducer/__init__.py
@@ -1,0 +1,15 @@
+from .reducer import at_least, avg, best_of, majority, median
+from .registry import create_reducers, reducer_log_names, score_reducer
+from .types import ScoreReducer
+
+__all__ = [
+    "ScoreReducer",
+    "score_reducer",
+    "create_reducers",
+    "reducer_log_names",
+    "avg",
+    "median",
+    "majority",
+    "best_of",
+    "at_least",
+]
