+++ src/inspect_ai/scorer/_reducer/types.py
@@ -1,0 +1,11 @@
+from typing import Protocol, runtime_checkable
+
+from .._metric import Score
+
+
+@runtime_checkable
+class ScoreReducer(Protocol):
+    def __call__(self, scores: list[Score]) -> Score: ...
+
+    @property
+    def __name__(self) -> str: ...
